Refactored implementation of the destination mock and Mock Launcher
